agentHost: give the injected MCP server names their own module
- `isHostInjectedMcpServerName` was parked in `claudeSdkOptions.ts`, which
  never used it, so customization discovery had to reach into the SDK-options
  module for a bare name check. Moving it to its own leaf module keeps the
  layering honest and leaves `claudeSdkOptions.ts` untouched by this change.
- `waitForAssert` takes an options bag instead of a positional
  `beforeAttempt`. Slotting a callback in ahead of the numeric parameters
  meant a future `waitForAssert(assertion, timeoutMs)` would silently be read
  as a callback; the bag removes that trap and leaves every existing call
  site as-is.

(Commit message generated by Copilot)
